[Integrate] Bump LLVM to 5912de9ede81407f93162e930ae9bc97e561d017 (#53714)
GPU_ORIG_REV_ID: 895039d2e6b184704bfec5731824c149f005d25b

+++ mojo/stdlib/src/gpu/mma.mojo
@@ -580,7 +580,7 @@ fn wgmma_commit_group_sync():
 @always_inline
 fn wgmma_wait_group_sync():
     __mlir_op.`nvvm.wgmma.wait.group.sync.aligned`[
-        _properties = __mlir_attr.`{group = 0 : i32}`, _type=None
+        _properties = __mlir_attr.`{group = 0 : i64}`, _type=None
     ]()
